\name{bcellNetwork}
\alias{bcellNetwork}
\docType{data}
\title{
B cell network
}
\description{
A subset of a molecular interaction network of Bcell assembled by the ARACNe\cite{Margolin2006} algorithm for protein-DN interactions and Bayesian method\cite{Lefebvre2010} for protein-
protein interactions.
}
\keyword{datasets}
